Moved Settings activity to prefs package
