+++ app/AndroidManifest.xml
@@ -48,7 +48,7 @@
         </activity>
 
         <activity
-            android:name="de.ironjan.mensaupb.activities.Settings_"
+            android:name="de.ironjan.mensaupb.prefs.Settings_"
             android:label="@string/app_name"
             android:parentActivityName="de.ironjan.mensaupb.activities.Menus_"
             android:theme="@style/TransparentActionBarActivity">
